Abstract factory calls now serialisable
Now use a container for dynamic template string tables
Factory parameters uses dynamic templates and is serialisable.
At the moment crates are saved to a different file using the new method, need to use an archive class so I can swap binary and text files
Need to put all of the other factorys in the level file.

Still need to convert all of the old factories, and then remove the old level file.